Rendering legacy vs refreshed Dashboard depending on privilege (#5298)
* Revert "Recent Forum messages in My Dashboard (#5282)"

This reverts commit b0afc5caf3b2ff19cd0e15050905d54c707ba206.

* Rendering legacy vs refreshed Dashboard depending on privilege

* Revert "Revert "Recent Forum messages in My Dashboard (#5282)""

This reverts commit 29f9f64452402ebca0851f9aa72e3b471cbff82b.
