+++ src/main/topLevelPages/Home/HomePage.tsx
@@ -1,34 +1,19 @@
-import { Box, Grow } from '@mui/material';
-import React, { useState } from 'react';
-import ContributorsSection from './ContributorsSection';
-import SpacesSection from '../../../domain/journey/space/DashboardSpaces/SpacesSection';
-import AnonymousUserHome from './AnonymousUserHome';
-import AuthenticatedUserHome from './AuthenticatedUserHome';
-import ReleaseUpdatesDialog from '../../../domain/platform/notifications/ReleaseUpdates/ReleaseUpdatesDialog';
+import React from 'react';
 import HomePageLayout from './HomePageLayout';
-import PageContent from '../../../core/ui/content/PageContent';
-import PageContentColumn from '../../../core/ui/content/PageContentColumn';
-import { gutters } from '../../../core/ui/grid/utils';
-import { useQueryParams } from '../../../core/routing/useQueryParams';
 import InnovationHubHomePage from '../../../domain/innovationHub/InnovationHubHomePage/InnovationHubHomePage';
 import Loading from '../../../core/ui/loading/Loading';
-import { useAuthenticationContext } from '../../../core/auth/authentication/hooks/useAuthenticationContext';
 import useInnovationHub from '../../../domain/innovationHub/useInnovationHub/useInnovationHub';
-import CreateAccountBanner from './CreateAccountBanner';
-import RecentJourneysList from '../myDashboard/recentJourneys/RecentJourneysList';
-import MyMembershipsDialog from '../myDashboard/myMemberships/MyMembershipsDialog';
-import MoreAboutAlkemio from '../myDashboard/moreAboutAlkemio/MoreAboutAlkemio';
-import RecentForumMessages from '../myDashboard/recentForumMessages/RecentForumMessages';
+import LegacyHomePage from '../myDashboard/legacy/LegacyHomePage';
+import useHasPlatformLevelPrivilege from '../../../domain/community/user/PlatformLevelAuthorization/useHasPlatformLevelPrivilege';
+import { AuthorizationPrivilege } from '../../../core/apollo/generated/graphql-schema';
+import MyDashboard from '../myDashboard/MyDashboard';
 
-export const HomePage = () => {
-  const { isAuthenticated, loading: isLoadingAuthentication } = useAuthenticationContext();
-
-  const params = useQueryParams();
-  const isFromLanding = params.get('from') === 'landing';
-
+const HomePage = () => {
   const { innovationHub, innovationHubLoading } = useInnovationHub();
 
-  const [isMyMembershipsDialogOpen, setIsMyMembershipsDialogOpen] = useState(false);
+  const [canAccessDashboardRefresh, { loading: isLoadingPrivileges }] = useHasPlatformLevelPrivilege(
+    AuthorizationPrivilege.AccessDashboardRefresh
+  );
 
   if (innovationHubLoading) {
     return (
@@ -42,29 +27,19 @@ export const HomePage = () => {
     return <InnovationHubHomePage innovationHub={innovationHub} />;
   }
 
-  return (
-    <HomePageLayout>
-      <ReleaseUpdatesDialog />
-      <MyMembershipsDialog open={isMyMembershipsDialogOpen} onClose={() => setIsMyMembershipsDialogOpen(false)} />
-      <PageContent>
-        <PageContentColumn columns={12}>
-          {!isLoadingAuthentication && !isAuthenticated && <CreateAccountBanner />}
-          <RecentJourneysList onSeeMore={() => setIsMyMembershipsDialogOpen(true)} />
-          {!isFromLanding && (
-            <Grow in={!isLoadingAuthentication} appear>
-              <Box display="flex" flexDirection="column" gap={gutters()} flexGrow={1} maxWidth="100%">
-                {isAuthenticated ? <AuthenticatedUserHome /> : <AnonymousUserHome />}
-              </Box>
-            </Grow>
-          )}
-          <SpacesSection />
-          <ContributorsSection />
-          <RecentForumMessages />
-          <MoreAboutAlkemio />
-        </PageContentColumn>
-      </PageContent>
-    </HomePageLayout>
-  );
+  if (isLoadingPrivileges) {
+    return (
+      <HomePageLayout>
+        <Loading />
+      </HomePageLayout>
+    );
+  }
+
+  if (canAccessDashboardRefresh) {
+    return <MyDashboard />;
+  } else {
+    return <LegacyHomePage />;
+  }
 };
 
 export default HomePage;
